Added alignment for allocations and unrolling.
Added alignment for allocations (defaulting to 16-bit aligned).

Added support for freeing previous allocation frames in the reverse
order
they were made. This is done by storing the size of an allocation in the
bytes at the top of an allocation frame. The size information can also
be used to traverse the allocation frames (which is now done when
dumping an arena).